PRSD-1108: Updates CheckLicensingAnswersPage to match design
