+++ src/main/resources/messages.properties
@@ -782,6 +782,8 @@ forms.update.occupancy.occupied.fieldSetHeading=Is your property still occupied 
 forms.update.numberOfHouseholds.fieldSetHeading=Update the number of households in the property
 forms.update.numberOfPeople.fieldSetHeading=Update how many people live in your property
 forms.update.licensingType.fieldSetHeading=Update the type of licensing you have for your property
+forms.update.checkLicensing.update.summaryName=You have updated the property licence
+forms.update.checkLicensing.remove.summaryName=You have removed this property\u2019s licence
 forms.update.warning=You must not provide any false or misleading information on the database. By submitting this information, you agree it\u2019s accurate to the best of your knowledge at the time of updating.
 
 forms.areYouSure.propertyDeregistration.fieldSetHeading=Are you sure you want to delete {0,,address} from the database?
